Move API response models to separate files.
